Addressing code review feedback - moving item constraints to assessment repo
