+++ service/src/main/java/tds/assessment/repositories/AssessmentQueryRepository.java
@@ -4,8 +4,7 @@ import java.util.List;
 import java.util.Optional;
 
 import tds.assessment.Assessment;
-import tds.assessment.Form;
-import tds.assessment.Item;
+import tds.assessment.ItemConstraint;
 
 /**
  * Queries for admin subject related entities
@@ -19,5 +18,14 @@ public interface AssessmentQueryRepository {
      * @return {@link tds.assessment.Assessment} if found otherwise empty
      */
     Optional<Assessment> findAssessmentByKey(String assessmentKey);
+
+    /**
+     * Finds a list of {@link tds.assessment.ItemConstraint}s for the assessment
+     *
+     * @param clientName
+     * @param assessmentId the assessment id for the item's {@link tds.assessment.Assessment}
+     * @return A list of item constraints for the assessment
+     */
+    List<ItemConstraint> findItemConstraintsForAssessment(final String clientName, final String assessmentId);
 
 }
