+++ service/src/main/java/tds/assessment/repositories/impl/ItemQueryRepositoryImpl.java
@@ -65,8 +65,6 @@ public class ItemQueryRepositoryImpl implements ItemQueryRepository {
                 }
         );
 
-
-
         return items;
     }
 
@@ -98,36 +96,4 @@ public class ItemQueryRepositoryImpl implements ItemQueryRepository {
         );
     }
 
-    @Override
-    public List<ItemConstraint> findItemConstraintsForAssessment(final String assessmentId) {
-        Map<String, Object> parameters = new HashMap<>();
-//        parameters.put("clientName", clientName);
-        parameters.put("assessmentId", assessmentId);
-
-        final String SQL =
-                "SELECT \n" +
-                        "   clientname,\n" +
-                        "   testid AS assessmentId, \n" +
-                        "   propname,\n" +
-                        "   propvalue,\n" +
-                        "   tooltype,\n" +
-                        "   toolvalue,\n" +
-                        "   item_in \n" +
-                        "FROM \n" +
-                        "   configs.client_test_itemconstraint \n" +
-                        "WHERE \n" +
-//                        "   clientname = :clientName AND \n" +
-                        "   testid = :assessmentId";
-
-        return jdbcTemplate.query(SQL, parameters, (rs, row) ->
-                new ItemConstraint.Builder()
-                        .withInclusive(rs.getBoolean("item_in"))
-                        .withPropertyName(rs.getString("propname"))
-                        .withPropertyValue(rs.getString("propvalue"))
-                        .withToolType(rs.getString("tooltype"))
-                        .withToolValue(rs.getString("toolvalue"))
-                        .withAssessmentId(rs.getString("assessmentId"))
-                        .build()
-        );
-    }
 }
