+++ service/src/main/java/tds/assessment/services/AssessmentService.java
@@ -12,8 +12,9 @@ public interface AssessmentService {
     /**
      * Retrieves a fully populated assessment
      *
+     * @param clientName the client environment identifier
      * @param assessmentKey to the assessment
      * @return {@link tds.assessment.Assessment} if found otherwise empty
      */
-    Optional<Assessment> findAssessmentByKey(String assessmentKey);
+    Optional<Assessment> findAssessmentByKey(String clientName, String assessmentKey);
 }
